benchmark/requirements.txt
gitpython>=3.1.0
unidiff>=0.7.0
pandas>=1.3.0
pyyaml
openai
azure-identity
python-dotenv>=0.21.0
azure-kusto-data